# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Erdos 服务端 API 契约
  version: 0.1.0-draft
  description: >
    客户端 / 引擎 / 服务端三方的共同语言。契约冻结项：错误码（x-error-codes）、
    表结构、签名格式——冻结后变更必须走契约评审（研发手册「质量标准」红线）。
    本文件为草案（x-contract-status: draft）：已 100% 覆盖实际实现（56 URL，2026-10-02
    契约一致性自检），待 ARCH 评审冻结为 v1。自检入口：server/scripts/contract_audit.py。
  x-contract-status: draft
servers:
  - url: https://api.erdos.example.com
...
paths:
...
  /v1/auth/login:
    post:
      tags: [auth]
      operationId: authLogin
      summary: 密码登录（签发 JWT 双令牌）
      description: >
        公开端点。校验通过签发访问令牌（Ed25519, 15 分钟）与刷新令牌（30 天）。
        连续失败达到阈值触发账号/IP 锁定（423），锁定与登录成功投影看板事件。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: 双令牌
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvelopeTokenPairView" }
        "401":
          description: 凭证无效（INVALID_CREDENTIALS）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
        "423":
          description: 账号已临时锁定（ACCOUNT_LOCKED）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
  /v1/auth/refresh:
    post:
      tags: [auth]
      operationId: authRefresh
      summary: 刷新轮换（旧刷新令牌即吊销）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "200":
          description: 新双令牌
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvelopeTokenPairView" }
        "401":
          description: 刷新令牌无效/已失效（TOKEN_REVOKED）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
  /v1/auth/logout:
    post:
      tags: [auth]
      operationId: authLogout
      summary: 注销（设备维度吊销，幂等）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LogoutRequest" }
      responses:
        "200":
          description: 吊销数量（已失效令牌返回 0）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvelopeLogoutView" }
...
  /v1/auth/register:
    post:
      tags: [auth]
      operationId: accountRegister
      summary: 注册（赠分领域事件 + 自动登录）
      description: >
        邮箱/手机至少其一；密码强度 ≥8 位含字母数字（40001）。
        同标识冲突 40901；同设备指纹重复注册不再触发赠分（防刷）。
        成功签发双令牌并发布 points.registration_grant 领域事件。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterCreate" }
      responses:
        "200":
          description: 双令牌（注册即登录）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvelopeTokenPairView" }
        "400":
          description: 参数/密码强度不满足（BAD_REQUEST）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
        "409":
          description: 邮箱或手机已注册（CONFLICT）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
  /v1/auth/password/reset/request:
    post:
      tags: [auth]
      operationId: passwordResetRequest
      summary: 申请密码重置（60s/次，每日 10 次）
      description: 无论账号是否存在均返回受理（防枚举）；令牌经渠道发送。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResetRequest" }
      responses:
        "200":
          description: 已受理
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvelopeResetAcceptedView" }
        "429":
          description: 发送过于频繁或当日限额（RATE_LIMITED）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
  /v1/auth/password/reset/confirm:
    post:
      tags: [auth]
      operationId: passwordResetConfirm
      summary: 确认密码重置（吊销全部会话）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResetConfirm" }
      responses:
        "200":
          description: 重置成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvelopeResetAcceptedView" }
        "400":
          description: 重置令牌无效或已过期（RESET_TOKEN_INVALID）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: JWT 双令牌（SP2-2 接入；SP2-1 网关仅做鉴权透传）
...
  schemas:
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: { type: string, minLength: 1, maxLength: 64 }
        password: { type: string, minLength: 1, maxLength: 128 }
        device_id: { type: string, maxLength: 64, nullable: true }
    RefreshRequest:
      type: object
      additionalProperties: false
      required: [refresh_token]
      properties:
        refresh_token: { type: string, minLength: 16, maxLength: 128 }
        device_id: { type: string, maxLength: 64, nullable: true }
    LogoutRequest:
      type: object
      additionalProperties: false
      required: [refresh_token]
      properties:
        refresh_token: { type: string, minLength: 16, maxLength: 128 }
    TokenPairView:
      type: object
      required: [access_token, refresh_token, token_type, expires_in, refresh_expires_in]
      properties:
        access_token: { type: string, description: Ed25519 签名访问令牌（15 分钟） }
        refresh_token: { type: string, description: 刷新令牌（30 天，轮换即吊销） }
        token_type: { type: string, example: bearer }
        expires_in: { type: integer, description: 访问令牌有效期（秒） }
        refresh_expires_in: { type: integer, description: 刷新令牌有效期（秒） }
    LogoutView:
      type: object
      required: [revoked]
      properties:
        revoked: { type: integer, description: 实际吊销的刷新令牌数量（幂等） }
    EnvelopeTokenPairView:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/TokenPairView" }
    EnvelopeLogoutView:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/LogoutView" }
    RegisterCreate:
      type: object
      additionalProperties: false
      required: [password, fingerprint]
      properties:
        email: { type: string, maxLength: 255, nullable: true }
        phone: { type: string, maxLength: 32, nullable: true }
        password: { type: string, minLength: 8, maxLength: 128 }
        fingerprint: { type: string, minLength: 8, maxLength: 64, description: 设备指纹（赠分防刷） }
        platform: { type: string, maxLength: 32, nullable: true }
    ResetRequest:
      type: object
      additionalProperties: false
      required: [identifier]
      properties:
        identifier: { type: string, minLength: 3, maxLength: 255, description: 邮箱或手机号 }
    ResetConfirm:
      type: object
      additionalProperties: false
      required: [reset_token, new_password]
      properties:
        reset_token: { type: string, minLength: 16, maxLength: 128 }
        new_password: { type: string, minLength: 8, maxLength: 128 }
    ResetAcceptedView:
      type: object
      required: [accepted]
      properties:
        accepted: { type: boolean }
    EnvelopeResetAcceptedView:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/ResetAcceptedView" }
...
    Envelope:
      type: object
      description: 统一响应信封。code=0 表示成功；其它取值与 x-error-codes 一一对应。
      required: [code, message, request_id, timestamp]
      properties:
        code:
          type: integer
          description: 业务码，0 表示成功，非 0 见 x-error-codes
        message:
          type: string
          description: 面向使用者的提示文案
        detail:
          type: string
          description: 可选的错误上下文（仅失败时携带）
          nullable: true
        data:
          description: 业务数据（成功时携带）
          nullable: true
        request_id:
          type: string
          description: 请求追踪 ID，与响应头 X-Request-Id 一致
        timestamp:
          type: string
          format: date-time
          description: 服务端 UTC 时间
...